Moved setup code to separate PHP file
